Update and move paginate test functions to utils package
